pyproject.toml
[tool.poetry]
name = "authentication-backend"
version = "0.1.0"
description = "Authentication API built by FastSurvey"
authors = ["Moritz Makowski <moritz@dostuffthatmatters.dev>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.61.1"
motor = "^2.3.0"
pymongo = "^3.11.0"
uvicorn = "^0.12.1"
certifi = "^2020.6.20"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
dnspython = "^2.0.0"
requests = "^2.24.0"
httpx = "^0.16.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pytest-cov = "^2.10.1"
autopep8 = "^1.5.4"
python-dotenv = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 80
ignore = "E501,W6"
